Update connectfour.c
This was what I was able to get done today.  I spent a lot of the beginning of the day getting reacquainted with C and planning this project out. This is as far as I got today. I plan to be pushing more frequently in the rest of the development of this project. The basics of this program are now set up. The user is prompted to input the size of the board they want, then greeted with a menu. Assuming they choose to play the game with a friend (solo play still needs to be developed), they are then brought into a match with a board they can populate with Xs and Os. The three major components that I need to add still are the win detection function(s), the AI, and the score keeping system. I plan to do as much as I can tomorrow. I will hopefully be able to complete the win detection and score keeping. I still need to hear back from Dr Anderson, as I posted on the forum making sure that the DFS algorithm I use for the win condition function is okay. This was a concern of mine since it will not function like a typical DFS because it must detect how many pieces are in a row, not just how far an extension is from the last placed piece. Hopefully she responds tomorrow and I can begin working on that function.