Update connectfour.c
I added around another 200 lines today. The only thing I have left to do is code the AI then clean up and organize a little bit of the code and possibly GUI. That will be done sometime in the next week or two. I am pretty proud of where the program is right now. I feel that I was able to do this pretty efficiently. I did not have nay major setbacks while programming this, and was able to get nearly the entire project finished in a matter or 3 days. I plan to spend a good bit of time working the AI out to try and make it competitive. From the rubric, this project can "borrow" 30% of it's code from other places, and I have yet to borrow any code. This means that I may re-purpose a C++ program that found the best next move for a game of connect 4. That may be too difficult however. I will most likely end up writing my own.