Added extra error checking
This update introduces Dr Lusth's functions to be used for handling user input so no extraneous input is allowed. Other bugs were found and tweaked as well. The AI was also started. I have a good idea on how to finish it. I want to work on it badly, but I have had so much school work and exams. I am finally free tomorrow afternoon and Thursday however, so I plan to finish the code up then.